app/requirements.txt
fastapi==0.115.0
langchain_core==0.3.12
langchain_groq==0.2.0
langgraph==0.2.24
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.0.8
uvicorn==0.30.6
langchain_openai==0.2.3